requirements.txt
click==8.1.3
easydict==1.10
numpy==1.18.1
open3d==0.17.0
pytorch_lightning==1.5.10
PyYAML==6.0
setuptools==59.5.0
spconv_cu113==2.3.6
tqdm==4.65.0
